Refactored document viewers into components
We will want to do further work on these to make them more reuseably, this is just an initial refactor to move them out of DocumentScreen.
