+++ ui/src/components/DocumentViewers/ImageViewer.jsx
