+++ ui/src/components/DocumentViewers/ImageViewer.scss
