+++ ui/src/components/DocumentViewers/PdfViewer.scss
