+++ ui/src/components/DocumentViewers/TextViewer.scss
